Heavy modifications to the newblog page
Eventually we will be able to drag elements from the 'Blog Elements'
sidebar to the center and they will be added to the dynamic form. Right
now the only one that adds is the textarea but when it adds it only
keeps one of the tabs open. The tab headers go invisible by design but
because the tabs are closing the whole thing looks incredibly broken,
which it is.